pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-pr-reviewer"
version = "0.1.0"
description = "LLM-powered GitHub Pull Request reviewer"
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "Your Name" }]
dependencies = [
  "openai>=1.47.0",
  "requests>=2.32.3",
  "PyYAML>=6.0.2",
  "tqdm>=4.66.5"
]

[project.scripts]
pr-reviewer = "pr_reviewer.main:cli"
